+++ eurostag-network/src/main/java/eu/itesla_project/eurostag/network/EsgDCNode.java
@@ -15,10 +15,10 @@ public class EsgDCNode {
 
     private final Esg2charName area; // zone identifier
     private final Esg8charName name; // node name
-    private final float vbase; // base voltage [kV]
-    private final float vinit; // Initial voltage [p.u.]
+    private final double vbase; // base voltage [kV]
+    private final double vinit; // Initial voltage [p.u.]
 
-    public EsgDCNode(Esg2charName area, Esg8charName name, float vbase, float vinit) {
+    public EsgDCNode(Esg2charName area, Esg8charName name, double vbase, double vinit) {
         this.area = Objects.requireNonNull(area);
         this.name = Objects.requireNonNull(name);
         this.vbase = vbase;
@@ -33,11 +33,11 @@ public class EsgDCNode {
         return name;
     }
 
-    public float getVbase() {
+    public double getVbase() {
         return vbase;
     }
 
-    public float getVinit() {
+    public double getVinit() {
         return vinit;
     }
